fix: separate the cbridge functions to each hooks
