+++ src/hooks/index.ts
@@ -15,5 +15,7 @@ export * from './useChainInfo';
 export * from './useConnectWallet';
 export * from './useUnbondWithdraw';
 export * from './useCustomSignature';
-export * from './useCbridge';
 export * from './useBreakpoints';
+export * from './c-bridge/useCbridge';
+export * from './c-bridge/useCbridgeHistory';
+export * from './c-bridge/useCbridgeApproval';
