+++ src/store/bridge/state.ts
@@ -1,0 +1,13 @@
+import { SelectedToken } from 'src/c-bridge';
+
+export interface GeneralStateInterface {
+  selectedToken: SelectedToken | null;
+}
+
+function state(): GeneralStateInterface {
+  return {
+    selectedToken: null,
+  };
+}
+
+export default state;
