+++ src/store/index.ts
@@ -9,6 +9,7 @@ import { DappStateInterface } from './dapp-staking/state';
 import general from './general';
 import contracts from './contracts';
 import dapps from './dapp-staking';
+import bridge from './bridge';
 
 /*
  * If not building with SSR mode, you can
@@ -44,6 +45,7 @@ export default store(function (/* { ssrContext } */) {
       general,
       contracts,
       dapps,
+      bridge,
     },
 
     // enable strict mode (adds overhead!)
